added API call for movie credits for individual actor ID
Realizing I'm pretty much going to be doing the meat of everything in
the success callbacks of each API method. If I do eventually want to
find a way to return data to my main jquery method it would be a hassle
so I'm just going to go forward doing it this way, feels like spaghetti
though.